chore: pull out components into folders
